Разрешил проблему проверки авторизации, принимаю обработку куки и тем самым удостоверяюсь что пользвоатель авторизован, добавил валидпацию запросов на форум

+++ packages/server/forum/controllers/commentController.ts
@@ -21,12 +21,14 @@ const getComments = (req: Request, res: Response) => {
 };
 const postComment = (req: Request, res: Response) => {
   const { message_id } = req.params;
+  const { user } = res.locals;
+  const { id, login, avatar } = user;
 
   CommentModel.create({
     ...req.body,
-    owner_avatar: 'TEST AVATAR',
-    owner_id: 'TEST ID',
-    owner_login: 'TEST LOGIN',
+    owner_avatar: avatar,
+    owner_id: id,
+    owner_login: login,
     message_id,
   })
     .then((comment) => {
@@ -38,37 +40,52 @@ const postComment = (req: Request, res: Response) => {
 };
 const updateComment = (req: Request, res: Response) => {
   const { id } = req.params;
+  const { user } = res.locals;
+  const { login, avatar } = user;
 
-  CommentModel.update(
-    {
-      ...req.body,
-      owner_avatar: 'SUPERTEST AVATAR',
-      owner_id: 'SUPERTEST ID',
-      owner_login: 'SUPERTEST LOGIN',
-    },
-    { where: { id }, returning: true },
-  )
-    .then((comment) => {
-      res.status(200).json(comment[1][0]);
-    })
-    .catch(() => {
-      res.status(400).json({ message: 'Bad request' });
-    });
+  CommentModel.findOne({ where: { id } }).then((comment) => {
+    if (comment && String(user.id) === comment.owner_id) {
+      CommentModel.update(
+        {
+          ...req.body,
+          owner_avatar: avatar,
+          owner_id: user.id,
+          owner_login: login,
+        },
+        { where: { id }, returning: true },
+      )
+        .then((comment) => {
+          res.status(200).json(comment[1][0]);
+        })
+        .catch(() => {
+          res.status(400).json({ message: 'Bad request' });
+        });
+    } else {
+      res.status(403).json({ message: 'Wrong owner' });
+    }
+  });
 };
 const deleteComment = (req: Request, res: Response) => {
   const { id } = req.params;
+  const { user } = res.locals;
 
-  CommentModel.destroy({
-    where: { id },
-  })
-    .then((deletedRecord) => {
-      deletedRecord === 1
-        ? res.status(200).json({ message: 'OK' })
-        : res.status(404).json({ message: 'Not found' });
-    })
-    .catch(() => {
-      res.status(400).json({ message: 'Bad request' });
-    });
+  CommentModel.findOne({ where: { id } }).then((comment) => {
+    if (comment && String(user.id) === comment.owner_id) {
+      CommentModel.destroy({
+        where: { id },
+      })
+        .then((deletedRecord) => {
+          deletedRecord === 1
+            ? res.status(200).json({ message: 'OK' })
+            : res.status(404).json({ message: 'Not found' });
+        })
+        .catch(() => {
+          res.status(400).json({ message: 'Bad request' });
+        });
+    } else {
+      res.status(403).json({ message: 'Wrong owner' });
+    }
+  });
 };
 
 export { getComments, postComment, updateComment, deleteComment };
